Remove the amIAdmin cloud function
Firestore rule protect the data against unauthorized access.
The function amIAdmin was used by the UI to determine if the currently logged
in user is an admin of a certain organization.
Being a serverless function it runs in a short lived environment that might need launch at
request time. So often the function would take one second to execute.

Using a query to firestore we can achieve the same result (determine if the current user
is an admin) with more predictable performance.
It's also a simpler solution with less code.