Use script path instead of current directory to determine esbuild config
The current directory might not always be the same: one might dare to run

    node functions/build_scripts/build.js

from the `eisbuk-admin` directory. We can't rely on `cwd` to be a specific one.
This change makes the build script rely on _its_ location instead of
the directory the command was run from, and thus is less brittle.