Fix BUILD_ENV condition
Production was broken. It had the __isDev__ flag set to `true`.
The reason for this was the condition inspecting the `BUILD_ENV`
variable did a strict comparison against the empty string:
it should be a dev build if and only if the `BUILD_ENV` variable
is different from the empty string.

But since we don't provide a value to the `BUILD_ENV` variable in the CI
job that does the deployment, `process.env.BUILD_ENV` ended up
having the value `undefined`, and a dev build was being pushed to production.

This commit changes the condition to check if the variable has a truthy value.
This should be what we want in this case: a non empty string will yield a
`true` while an `undefined` or an empty string will yield a `false`.
See https://developer.mozilla.org/en-US/docs/Glossary/Truthy