Work around difficulty specifying a source directory that is not `src`
The esbuild bundler will look into the given `entryPoints`.
If we pass a path like `instrumented/index.tsx` it will still load
all other files from `src/`.
This needs a better solution, but for now we can bait-and-switch
the `instrumented` directory in place of the `src` one during the build.