Remove unnecessary click statement
This test needs to wait for slots to be loaded before clicking on the "copy" button.
Now it does this explicitly (by waiting for the word "competitive" to appear on the page)
but it was not always the case.

Before the explicit wait was added, the waiting was sort-of there, since the two click
statemts _usually_ took enough time for the slots to load in the background.

Now that the explicit wait is in there's no need anymore for the additional click.